Retrieve standardized book names based on biblical authors

Collect book names from RASB Bible, Leningrad Codex, or Septuagint based on biblical authors outlined in 'author_data' dataset.

Moses: Genesis – Deuteronomy

Joshua: Joshua

Unkown: Judges, Ruth, 1 & 2 Kings, 1 & Chronicles, Esther and Job

Samuel/Unknown: 1 & 2 Samuel

Ezra: Ezra

Nehemiah: Nehemiah

David/Various: Psalms

Solomon: Proverbs, Ecclesiastes, Song of Solomon

Isaiah: Isaiah

Jeremiah: Jeremiah and Lamentations

Ezekiel: Ezekiel

Daniel: Daniel

Hosea: Hosea

Joel: Joel

Amos: Amos

Obadiah: Obadiah

Jonah: Jonah

Micah: Micah

Nahum: Nahum

Habakkuk: Habakkuk

Zephaniah: Zephaniah

Haggai: Haggai

Zechariah: Zechariah

Malachi: Malachi

Matthew: Gospel of Matthew

Mark: Gospel of Mark

Luke: Gospel of Luke and Acts

John: Gospel of John, Johanine epistles and Revelation

Paul: Pauline epistles

James: James

Peter: 1 and 2 Peter

Jude: Jude

Unknown: Hebrews

Usage

by_author(author, testament)

Arguments

author

a character vector or length 1 specifying the biblical author in the Old or New Testament

testament

a character vector of length 1 specifying the testament/time period in which the author is found

Value

a character vector of standardized book names to be passed to other functions