Retrieve standardized book names based on biblical narrative styles or genres

Collect book names from RASB Bible, Leningrad Codex, or Septuagint based on narrative styles outlined in 'author_data' dataset.

Law: Genesis – Deuteronomy

History: Joshua – Esther

Wisdom: Job – Songs of Solomon

Major Prophets: Isaiah - Daniel

Minor Prophets: Hosea – Malachi

Gospels: Matthew – John

Pauline Epistles: Romans – Philemon

General Epistles: Hebrews – Jude

Apocalyptic: Revelation

Usage

by_section(section, testament)

Arguments

section

a character vector or length 1 specifying the narrative section or genre in the Old or New Testament

testament

a character vector of length 1 specifying the testament in which the section is found

Value

a character vector of standardized book names to be passed to other functions