Retrieve a specific chapter or portion of a chapter from the Bible

This function retrieves text from a specified book and chapter, with optional filtering by verse, partitioning, and language selection.

Usage

retrieve_chapter(
  book,
  chapter = NULL,
  verse = NULL,
  fraction = NULL,
  part = NULL,
  language,
  testament
)

Arguments

book

A character string specifying the book of the Bible.

chapter

A numeric or character vector specifying the chapter(s) to retrieve.

verse

An optional numeric vector specifying specific verses to retrieve.

fraction

A numeric value indicating how many equal parts to divide the chapter into.

part

A numeric value specifying which part to return (must be between 1 and fraction).

language

A character string specifying the language of the Bible text. Options are "English", "Hebrew", or "Greek".

testament

A character string specifying whether to retrieve from the Old or New Testament.

Value

A character vector containing the retrieved Bible text.

Examples

retrieve_chapter("Jud",chapter = 1, verse = 1, language = "English", testament = "new")
#> Did you mean: Jde ?
#> [1] "Jude, a servant of Jesus Christ, and brother of James, to them that are called, beloved in God the Father, and kept for Jesus Christ:"